+++ plugins/map.js
@@ -1,4 +1,6 @@
 import Vue from 'vue'
-import Vue2LeafletTracksymbol from 'vue2-leaflet-tracksymbol'
+import Vue2LeafletMarkerCluster from 'vue2-leaflet-markercluster' // Doc: https://github.com/Leaflet/Leaflet.markercluster
+// import Vue2LeafletTracksymbol from 'vue2-leaflet-tracksymbol' // Doc: https://github.com/lethexa/leaflet-tracksymbol
 
-Vue.component('l-ais', Vue2LeafletTracksymbol)
+Vue.component('l-marker-cluster', Vue2LeafletMarkerCluster)
+// Vue.component('l-ais', Vue2LeafletTracksymbol)